rename files, implementing handling both string and number as a value
